feat(api): split users contorller into action files with auth middleware
