+++ src/api/cms/user/actions/activateUser.ts
@@ -1,0 +1,26 @@
+import { FastifyInstance, RouteShorthandOptions } from "fastify"
+import { schemaRef } from "~/schema/cms-api"
+
+export default function activateUser(instance: FastifyInstance) {
+  instance.get<{
+    Params: { id: string }
+    Querystring: { hash: string }
+  }>("/users/:id/activate", activateOptions, async (req, reply) => {
+    const id = req.params.id
+
+    const result = await instance.users.selfActivate(id, req.query.hash)
+
+    if (!result) {
+      reply.code(404)
+    }
+
+    reply.send()
+  })
+}
+
+const activateOptions: RouteShorthandOptions = {
+  schema: {
+    params: schemaRef("idParamStrict"),
+    querystring: schemaRef("hashQueryStrict"),
+  },
+}
